overrides/Gio/ListStore: Add type annotations
Since we are already defining a type variable for the Generic parent class,
we might as well put it to good use and add type annotations to the methods.

These do not really appear to do anything other than serve as a documentation for now.

Had to disable E704 flake8 rule since it misplaces ellipsis.
In fact, this is why it is ignored in the default configuration:
https://github.com/PyCQA/pycodestyle/issues/973