Support instances of Generic types in propertyhelper
Subscribing a subclass of `Generic[...]` (e.g. `Gio.ListStore[Gio.File]`)
will turn it into instance of `typing._GenericAlias`.
That is not an instance of `type`, nor a subclass of the subscribed class.

To allow passing instantiated generic types as `type` argument of `Property`,
we need to extract the generic type, or property will fail with `Unsupported type` error.

Unfortunately, `typing` module only provides helper for that since Python 3.8
so we need to create our own `get_origin` that looks into internal state for 3.7.